Feat: 음악 장르 매핑 테이블 추가

+++ src/main/java/com/mate/band/domain/user/dto/RegisterProfileRequestDTO.java
@@ -37,7 +37,7 @@ public record RegisterProfileRequestDTO(
         List<ProfileSNSUrlDTO> snsUrls,
 
         @Schema(description = "합주 가능 지역 코드", example = "[1, 2, 3]")
-        List<Integer> district,
+        List<Long> district,
 
         @Schema(description = "선호 장르 코드", example = "[\"kpop\", \"indie\", \"jazz\"]")
         List<String> genre,
